edicion listo

+++ database/migrations/2020_11_11_021232_create_refaccionarias_table.php
@@ -18,7 +18,7 @@ class CreateRefaccionariasTable extends Migration
             $table->string('nombre');
             $table->string('descripcion');
             $table->integer('numPiezas');
-            $table->integer('costoPieza');
+            $table->float('costoPieza');
             $table->timestamps();
         });
     }
